% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mschart.R
\docType{data}
\name{browser_data}
\alias{browser_data}
\title{Dummy dataset for barchart}
\format{A data frame with 18 rows and 3 variables}
\usage{
data(browser_data)
}
\description{
A dataset containing 2 categorical and an integer variables:
}
\details{
\itemize{
  \item browser web browser
  \item serie id of series
  \item value integer values
}
}
\keyword{datasets}
