% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{browser_ts}
\alias{browser_ts}
\title{Dummy dataset for time series charts}
\format{
A data frame with 36 rows and 3 variables
}
\description{
A dataset containing a date, a categorical, and an integer variable:
}
\details{
\itemize{
\item date date values
\item browser web browser
\item freq values in percent
}
}
