% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_setting.R
\name{chart_ax_x}
\alias{chart_ax_x}
\alias{chart_ax_x.ms_chart}
\title{x axis settings}
\usage{
chart_ax_x(x, orientation, crosses, cross_between, major_tick_mark,
  minor_tick_mark, tick_label_pos, display, num_fmt, rotation, limit_min,
  limit_max, position, second_axis = FALSE)

\method{chart_ax_x}{ms_chart}(x, orientation, crosses, cross_between,
  major_tick_mark, minor_tick_mark, tick_label_pos, display, num_fmt, rotation,
  limit_min, limit_max, position, second_axis = FALSE)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{orientation}{axis orientation, one of 'maxMin', 'minMax'.}

\item{crosses}{specifies how the axis crosses the perpendicular
axis, one of 'autoZero', 'max', 'min'.}

\item{cross_between}{specifies how the value axis crosses the
category axis between categories, one of 'between', 'midCat'.}

\item{major_tick_mark, minor_tick_mark}{tick marks position,
one of 'cross', 'in', 'none', 'out'.}

\item{tick_label_pos}{ticks labels position, one of 'high',
'low', 'nextTo', 'none'.}

\item{display}{should the axis be displayed (a logical of
length 1).}

\item{num_fmt}{number formatting. It can be "General", "0.00",
"#,##0", "#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{rotation}{rotation angle. Value should be between `-360`
and `360`.}

\item{limit_min}{minimum value on the axis.}

\item{limit_max}{maximum value on the axis.}

\item{position}{position value that cross the other axis.}

\item{second_axis}{unused}
}
\description{
Define settings for an x axis.
}
\section{Methods (by class)}{
\itemize{
\item \code{ms_chart}: chart_ax_x method for ms_chart objects
}}

\seealso{
\code{\link{chart_ax_y}}
}
