% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_setting.R
\name{chart_ax_y}
\alias{chart_ax_y}
\alias{chart_ax_y.ms_chart}
\title{y axis settings}
\usage{
chart_ax_y(x, orientation, crosses, cross_between, major_tick_mark,
  minor_tick_mark, tick_label_pos, display, num_fmt, rotation, limit_min,
  limit_max, position, second_axis = FALSE)

\method{chart_ax_y}{ms_chart}(x, orientation, crosses, cross_between,
  major_tick_mark, minor_tick_mark, tick_label_pos, display, num_fmt, rotation,
  limit_min, limit_max, position, second_axis = FALSE)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{orientation}{`character(1)`: axis orientation,
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_orientation)}.}

\item{crosses}{`character(1)`: specifies how the axis crosses the perpendicular axis,
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_crosses)}.}

\item{cross_between}{`character(1)`: specifies how the value axis crosses the category
axis between categories,
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_crossbetween)}.}

\item{major_tick_mark}{`character(1)`: tick marks position,
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_tickmark)}.}

\item{minor_tick_mark}{`character(1)`: tick marks position,
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_tickmark)}.}

\item{tick_label_pos}{`character(1)`: ticks labels position,
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_ticklblpos)}.}

\item{display}{`logical(1)`: should the axis be displayed.}

\item{num_fmt}{`character(1)`: number formatting. It can be "General", "0.00", "#,##0",
"#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{rotation}{`integer(1)`: rotation angle. Value should be between `-360` and `360`.}

\item{limit_min}{`numeric(1)`: minimum value on the axis.}

\item{limit_max}{`numeric(1)`: maximum value on the axis.}

\item{position}{`numeric(1)`: position value that cross the other axis.}

\item{second_axis}{`logical(1)`: unused}
}
\description{
Set y axis properties.
}
\section{Methods (by class)}{
\itemize{
\item \code{ms_chart}: chart_ax_y method for ms_chart objects
}}

\seealso{
\code{\link{chart_ax_x}}
}
