% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_color.R
\name{chart_data_fill}
\alias{chart_data_fill}
\title{Modify fill colour}
\usage{
chart_data_fill(x, values)
}
\arguments{
\item{x}{chart object}

\item{values}{`character(num of series)`: a set of colours values to map data values to.
It is a named vector, the values will be matched based on the names.}
}
\description{
Modify fill colour for series
}
