% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_labels_options.R
\name{chart_data_labels}
\alias{chart_data_labels}
\title{Modify data labels settings}
\usage{
chart_data_labels(x, num_fmt = "General", position = "ctr",
  show_legend_key = FALSE, show_val = FALSE, show_cat_name = FALSE,
  show_serie_name = FALSE, show_percent = FALSE, separator = ", ")
}
\arguments{
\item{x}{chart object}

\item{num_fmt}{`character(1)`: number formatting specifies number format properties which
indicate how to format and render the numeric values. It can be "General", "0.00", "#,##0",
"#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{position}{`character(1)`: it specifies the position of the data label. It should be
one of \Sexpr[stage=render, results=rd]{mschart:::choices_rd(mschart:::st_dlblpos)}}

\item{show_legend_key}{`logical(1)`}

\item{show_val}{`logical(1)`}

\item{show_cat_name}{`logical(1)`}

\item{show_serie_name}{`logical(1)`}

\item{show_percent}{`logical(1)`}

\item{separator}{`character(1)`}
}
\description{
Data labels show details about data series.
}
