% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_labels.R
\name{chart_data_labels_override}
\alias{chart_data_labels_override}
\title{Override data labels settings}
\usage{
chart_data_labels_override(x, values)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{values}{a named list of lists of \code{\link[=labels_options]{labels_options()}} objects or \code{NA}s to map data labels to.
It is a named list, the values will be matched based on the names.
Values are mapped to data points by position. Each list must match the respective series' length.
\code{NA} means that the series-level setting will be used.}
}
\description{
Specify data point specific mappings from levels in the data to displayed data labels.
}
\examples{
library(officer)

label_settings <- list(
  serie1 = labels_options(show_val = TRUE),
  serie2 = labels_options(show_cat_name = TRUE),
  serie3 = labels_options(show_percent = TRUE)
)

label_settings_override <- list(
  serie1 = list(NA, labels_options(show_val = TRUE, show_legend_key = TRUE), NA, NA, NA, NA)
)

barchart <- ms_barchart(
  data = browser_data,
  x = "browser", y = "value", group = "serie")
barchart <- chart_data_labels(barchart, show_val = TRUE)
barchart <- chart_data_labels_series( barchart,
  values = label_settings )
barchart <- chart_data_labels_override(barchart, values = label_settings_override)
}
\seealso{
\code{\link[=chart_data_labels]{chart_data_labels()}} \code{\link[=chart_data_labels_series]{chart_data_labels_series()}}

Other Data point customization functions: 
\code{\link{chart_data_fill_override}()},
\code{\link{chart_data_line_style_override}()},
\code{\link{chart_data_line_width_override}()},
\code{\link{chart_data_size_override}()},
\code{\link{chart_data_stroke_override}()},
\code{\link{chart_data_symbol_override}()},
\code{\link{chart_labels_text_override}()}
}
\concept{Data point customization functions}
