% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_labels.R
\name{chart_data_labels_series}
\alias{chart_data_labels_series}
\title{Modify data labels settings}
\usage{
chart_data_labels_series(x, values = NULL)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{values}{a named list of \code{\link[=labels_options]{labels_options()}} objects to map data labels to.
It is a named list, the values will be matched based on the names.
If it contains only one \code{\link[=labels_options]{labels_options()}} object, it will be associated to all existing series.}
}
\description{
Specify mappings from levels in the data to displayed data labels.
Overrides settings from \code{\link[=chart_data_labels]{chart_data_labels()}}.
}
\examples{
library(officer)

label_settings <- list(
  serie1 = labels_options(show_val = TRUE),
  serie2 = labels_options(show_cat_name = TRUE),
  serie3 = labels_options(show_percent = TRUE)
)

barchart <- ms_barchart(
  data = browser_data,
  x = "browser", y = "value", group = "serie")
barchart <- chart_data_labels(barchart, show_val = TRUE)
barchart <- chart_data_labels_series( barchart,
  values = label_settings )
}
\seealso{
Other Series customization functions: 
\code{\link{chart_data_fill}()},
\code{\link{chart_data_line_style}()},
\code{\link{chart_data_line_width}()},
\code{\link{chart_data_size}()},
\code{\link{chart_data_smooth}()},
\code{\link{chart_data_stroke}()},
\code{\link{chart_data_symbol}()},
\code{\link{chart_labels_text}()}
}
\concept{Series customization functions}
