% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_styles.R
\name{chart_data_size_override}
\alias{chart_data_size_override}
\title{Override symbol size}
\usage{
chart_data_size_override(x, values)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{values}{a named list of vectors of size values or \code{NA}s to map data values to.
It is a named list, the values will be matched based on the names.
Values are mapped to data points by position. Each list must match the respective series' length.
\code{NA} means that the series-level setting will be used.}
}
\description{
Specify data point specific mappings from levels in the data to displayed size of symbols.
}
\examples{
my_scatter <- ms_scatterchart(data = iris, x = "Sepal.Length",
  y = "Sepal.Width",  group = "Species")
my_scatter <- chart_data_fill(my_scatter,
  values = c(virginica = "#6FA2FF", versicolor = "#FF6161", setosa = "#81FF5B") )
my_scatter <- chart_data_stroke(my_scatter,
  values = c(virginica = "black", versicolor = "black", setosa = "black") )
my_scatter <- chart_data_size(my_scatter,
  values = c(virginica = 20, versicolor = 16, setosa = 20) )

my_scatter <- chart_data_size_override(my_scatter,
  values = list(virginica = c(30, rep(NA, 149))))
}
\seealso{
\code{\link[=chart_data_size]{chart_data_size()}}

Other Data point customization functions: 
\code{\link{chart_data_fill_override}()},
\code{\link{chart_data_labels_override}()},
\code{\link{chart_data_line_style_override}()},
\code{\link{chart_data_line_width_override}()},
\code{\link{chart_data_stroke_override}()},
\code{\link{chart_data_symbol_override}()},
\code{\link{chart_labels_text_override}()}
}
\concept{Data point customization functions}
