% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_styles.R
\name{chart_data_smooth}
\alias{chart_data_smooth}
\title{Smooth series}
\usage{
chart_data_smooth(x, values)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{values}{\verb{integer(num of series)}: a set of smooth values to map data values to.
It is a named vector, the values will be matched based on the names.
Use \code{0} to disable smoothing and \code{1} to enable it.
If it contains only one integer it will be associated to all existing series.}
}
\value{
An \code{ms_chart} object.
}
\description{
Specify whether lines should be smoothed, per series. This
feature only applies to \code{\link[=ms_linechart]{ms_linechart()}}.
}
\examples{
linec <- ms_linechart(data = iris, x = "Sepal.Length",
  y = "Sepal.Width", group = "Species")
linec <- chart_data_smooth(linec,
  values = c(virginica = 0, versicolor = 0, setosa = 0) )
}
\seealso{
Other Series customization functions: 
\code{\link{chart_data_fill}()},
\code{\link{chart_data_line_style}()},
\code{\link{chart_data_line_width}()},
\code{\link{chart_data_size}()},
\code{\link{chart_data_stroke}()},
\code{\link{chart_data_symbol}()},
\code{\link{chart_labels_text}()}
}
\concept{Series customization functions}
