% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_color.R
\name{chart_data_symbol}
\alias{chart_data_symbol}
\title{Modify symbol}
\usage{
chart_data_symbol(x, values)
}
\arguments{
\item{x}{chart object}

\item{values}{`character(num of series)`: a set of symbol values to map data values to.
It is a named vector, the values will be matched based on the names.}
}
\description{
Modify symbol for series
}
