% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_styles.R
\name{chart_labels_text_override}
\alias{chart_labels_text_override}
\title{Override labels font settings}
\usage{
chart_labels_text_override(x, values)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{values}{a named list of lists of \code{\link[=fp_text]{fp_text()}} objects or \code{NA}s to map data labels to.
It is a named list, the values will be matched based on the names.
Values are mapped to data points by position. Each list must match the respective series' length.
\code{NA} means that the series-level setting will be used.}
}
\description{
Specify data point specific mappings from levels in the data to displayed text font settings.
}
\examples{
library(officer)

fp_text_settings <- list(
  serie1 = fp_text(font.size = 7, color = "red"),
  serie2 = fp_text(font.size = 0, color = "purple"),
  serie3 = fp_text(font.size = 19, color = "wheat")
)

fp_text_settings_override <- list(
  serie1 = list(NA, NA, fp_text(font.size = 7, color = "blue"), NA, NA, NA)
)

barchart <- ms_barchart(
  data = browser_data,
  x = "browser", y = "value", group = "serie")
barchart <- chart_data_labels(barchart, show_val = TRUE)
barchart <- chart_labels_text( barchart,
  values = fp_text_settings )
barchart <- chart_labels_text_override(barchart, values = fp_text_settings_override)
}
\seealso{
\code{\link[=chart_labels_text]{chart_labels_text()}}

Other Data point customization functions: 
\code{\link{chart_data_fill_override}()},
\code{\link{chart_data_labels_override}()},
\code{\link{chart_data_line_style_override}()},
\code{\link{chart_data_line_width_override}()},
\code{\link{chart_data_size_override}()},
\code{\link{chart_data_stroke_override}()},
\code{\link{chart_data_symbol_override}()}
}
\concept{Data point customization functions}
