% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_setting.R
\name{chart_table}
\alias{chart_table}
\title{x table settings}
\usage{
chart_table(x, horizontal, vertical, outline, show_keys)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{horizontal}{write horizontal lines in the table}

\item{vertical}{write vertical lines in the table}

\item{outline}{write an outline in the table}

\item{show_keys}{showkeys in the table}
}
\description{
Define settings for an x table.
}
