% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{chart_theme}
\alias{chart_theme}
\title{Modify components of an \code{ms_chart} theme.}
\usage{
chart_theme(axis.title = fp_text(bold = TRUE, font.size = 16),
  axis.title.x = axis.title, axis.title.y = axis.title,
  axis.text = fp_text(), axis.text.x = axis.text, axis.text.y = axis.text,
  axis.ticks = fp_border(color = "#99999999"), axis.ticks.x = axis.ticks,
  axis.ticks.y = axis.ticks, grid.major.line = fp_border(color =
  "#99999999", style = "dashed"), grid.major.line.x = grid.major.line,
  grid.major.line.y = grid.major.line, grid.minor.line = fp_border(width =
  0), grid.minor.line.x = grid.minor.line,
  grid.minor.line.y = grid.minor.line, date_fmt = "yyyy/mm/dd",
  str_fmt = "General", double_fmt = "#,##0.00", integer_fmt = "0")
}
\arguments{
\item{axis.title, axis.title.x, axis.title.y}{axis title formatting properties (\link[officer]{fp_text})}

\item{axis.text, axis.text.x, axis.text.y}{axis text formatting properties (\link[officer]{fp_text})}

\item{axis.ticks, axis.ticks.x, axis.ticks.y}{axis ticks formatting properties (\link[officer]{fp_border})}

\item{grid.major.line, grid.major.line.x, grid.major.line.y, grid.minor.line, grid.minor.line.x, grid.minor.line.y}{grid lines formatting properties (\link[officer]{fp_border})}

\item{date_fmt}{date format}

\item{str_fmt}{string or factor format}

\item{double_fmt}{double format}

\item{integer_fmt}{integer format}
}
\description{
Use \code{chart_theme()} to modify individual components of a chart theme.
}
