% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_barchart.R
\name{ms_barchart}
\alias{ms_barchart}
\title{barchart object}
\usage{
ms_barchart(data, x, y, group = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{x}{x colname}

\item{y}{y colname}

\item{group}{grouping colname used to split data into series. Optional.}
}
\description{
Creation of a barchart object that can be
inserted in a 'Microsoft' document.
}
