% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{ms_combochart}
\alias{ms_combochart}
\title{combochart object}
\usage{
ms_combochart(...)
}
\arguments{
\item{...}{mschart objects}
}
\description{
Creation of a combochart object that can be
inserted in a 'Microsoft' document.
}
\details{
ms_combochart only works with mschart objects created with
\code{asis = TRUE}.
}
\examples{
# example chart 01 -------

dat <- longley
dat$Year <- as.Date(paste0(dat$Year, "-01-01"))
dat$`GNP Deflator` <- dat$GNP.deflator

empl <- ms_barchart(data = dat, x = "Year", y = c("Employed", "Unemployed"), asis = TRUE)
empl <- chart_labels(x = empl, title = "Longley's Economic Regression Data",
                     xlab = "1947 to 1962", ylab = "Employment")
empl <- chart_ax_x(x = empl, num_fmt = "yyyy")
empl <- as_bar_stack(x = empl, dir = "vertical")

gdp <- ms_linechart(data = dat, x = "Year", y = "GNP Deflator", asis = TRUE)
gdp <- chart_labels(x = gdp, ylab = "GNP implicit price deflator (1954 = 100)")
gdp <- chart_ax_y(x = gdp, second_axis = TRUE)

cb <- ms_combochart(empl, gdp)
}
\seealso{
\code{\link[=chart_settings]{chart_settings()}}, \code{\link[=chart_ax_x]{chart_ax_x()}}, \code{\link[=chart_ax_y]{chart_ax_y()}},
\code{\link[=chart_data_labels]{chart_data_labels()}}, \code{\link[=chart_theme]{chart_theme()}}, \code{\link[=chart_labels]{chart_labels()}}

Other 'Office' chart objects: 
\code{\link{ms_areachart}()},
\code{\link{ms_barchart}()},
\code{\link{ms_linechart}()},
\code{\link{ms_piechart}()},
\code{\link{ms_scatterchart}()}
}
\concept{'Office' chart objects}
