% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_scatter.R
\name{ms_scatter}
\alias{ms_scatter}
\title{scatter object}
\usage{
ms_scatter(data, x, y, group = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{x}{x colname}

\item{y}{y colname}

\item{group}{grouping colname used to split data into series. Optional.}
}
\description{
Creation of a scatter object that can be
inserted in a 'Microsoft' document.
}
