% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{ms_scatterchart}
\alias{ms_scatterchart}
\title{scatterchart object}
\usage{
ms_scatterchart(data, x, y, group = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{x}{x colname}

\item{y}{y colname}

\item{group}{grouping colname used to split data into series. Optional.}
}
\description{
Creation of a scatterchart object that can be
inserted in a 'Microsoft' document.
}
\examples{
library(officer)
# example chart_01 -------
chart_01 <- ms_scatterchart(data = mtcars, x = "disp",
                         y = "drat")
chart_01 <- chart_ax_x(chart_01, cross_between = "midCat")
chart_01 <- chart_settings(chart_01, scatterstyle = "marker")


}
\seealso{
Other 'Office' chart objects: 
\code{\link{ms_areachart}()},
\code{\link{ms_barchart}()},
\code{\link{ms_linechart}()}
}
\concept{'Office' chart objects}
