% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{mschart-defunct}
\alias{mschart-defunct}
\alias{ph_with_chart}
\alias{ph_with_chart_at}
\title{Defunct Functions in Package mschart}
\usage{
ph_with_chart(...)

ph_with_chart_at(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package mschart
}
\details{
\code{ph_with_chart()} is replaced by \code{ph_with.ms_chart}.

\code{ph_with_chart_at()} is replaced by \code{ph_with.ms_chart}.
}
