% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_with_chart.R
\name{ph_with_chart}
\alias{ph_with_chart}
\alias{ph_with_chart_at}
\title{add chart into a PowerPoint slide}
\usage{
ph_with_chart(x, value, type = "body", index = 1)

ph_with_chart_at(x, value, left, top, width, height)
}
\arguments{
\item{x}{a pptx device}

\item{value}{\code{ms_chart} object}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}

\item{left, top}{location of chart on the slide}

\item{height, width}{Height and width in inches.}
}
\description{
add a chart as a new shape in the current slide.
}
