% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{print.ms_chart}
\alias{print.ms_chart}
\title{ms_chart print method}
\usage{
\method{print}{ms_chart}(x, preview = FALSE, ...)
}
\arguments{
\item{x}{ms_chart object}

\item{preview}{preview the chart in a PowerPoint document}

\item{...}{unused}
}
\description{
ms_chart print method
}
