% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_setting.R
\name{set_x_axis}
\alias{set_x_axis}
\alias{set_y_axis}
\alias{set_x_axis.ms_chart}
\title{axis settings}
\usage{
set_x_axis(x, orientation, crosses, cross_between, major_tick_mark,
  minor_tick_mark, tick_label_pos, display, num_fmt, rotation,
  second_axis = FALSE)

set_y_axis(x, orientation, crosses, cross_between, major_tick_mark,
  minor_tick_mark, tick_label_pos, display, num_fmt, rotation,
  second_axis = FALSE)

\method{set_x_axis}{ms_chart}(x, orientation, crosses, cross_between,
  major_tick_mark, minor_tick_mark, tick_label_pos, display, num_fmt, rotation,
  second_axis = FALSE)
}
\arguments{
\item{x}{chart object}

\item{orientation}{axis orientation, expected value is one
of "maxMin" or "minMax".}

\item{crosses}{specifies how the axis crosses the perpendicular axis,
expected value should be one of "autoZero", "max" or "min".}

\item{cross_between}{specifies how the value axis crosses the category
axis between categories, expected value should be one of "midCat" or "between".}

\item{major_tick_mark, minor_tick_mark}{tick marks position, expected value should
be one of 'cross', 'in', 'none' or 'out'.}

\item{tick_label_pos}{ticks labels position, expected value should
be one of 'high', 'low', 'nextTo' or 'none'.}

\item{display}{should the axis be displayed.}

\item{num_fmt}{number formatting. It can be "General", "0.00", "#,##0",
"#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{rotation}{rotation angle.}

\item{second_axis}{unused}
}
\description{
Set axis properties.
}
