% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_codes.R
\name{table_content_xml}
\alias{table_content_xml}
\title{xml code for table in chart}
\usage{
table_content_xml(x)
}
\arguments{
\item{x}{an object of class table_options}
}
\description{
xml code for table in chart
}
