% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{us_indus_prod}
\alias{us_indus_prod}
\title{Index of US Industrial Production}
\format{
A data frame with 256 rows and 3 variables
}
\description{
Index of US industrial production (1985 = 100).
}
\details{
This is a transformation of the \code{USProdIndex} data
from package 'AER' into a simple data.frame.
}
