% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{browser_data}
\alias{browser_data}
\title{Dummy dataset for barchart}
\format{
A data frame with 18 rows and 3 variables
}
\description{
A dataset containing two categorical variables and one integer variable:
}
\details{
\itemize{
\item browser web browser
\item serie id of series
\item value integer values
}
}
