% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mschart.R
\docType{data}
\name{browser_ts}
\alias{browser_ts}
\title{Dummy dataset for barchart}
\format{A data frame with 36 rows and 3 variables}
\usage{
data(browser_ts)
}
\description{
A dataset containing a date, a categorical and an integer variables:
}
\details{
\itemize{
  \item date date values
  \item browser web browser
  \item freq values in percent
}
}
\keyword{datasets}
