% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_labels.R
\name{chart_data_labels}
\alias{chart_data_labels}
\title{Modify data labels settings}
\usage{
chart_data_labels(x, num_fmt = "General", position = "ctr",
  show_legend_key = FALSE, show_val = FALSE, show_cat_name = FALSE,
  show_serie_name = FALSE, show_percent = FALSE, separator = ", ")
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{num_fmt}{\code{character(1)}: number formatting specifies number format properties which
indicate how to format and render the numeric values. It can be "General", "0.00", "#,##0",
"#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{position}{\code{character(1)}: it specifies the position of the data label.
It should be one of 'b', 'ctr', 'inBase', 'inEnd', 'l',
'outEnd', 'r', 't'.
When grouping is 'clustered', it should be one of 'ctr','inBase','inEnd','outEnd'.
When grouping is 'stacked', it should be one of 'ctr','inBase','inEnd'.
When grouping is 'standard', it should be one of 'b','ctr','l','r','t'.}

\item{show_legend_key}{show legend key if TRUE.}

\item{show_val}{show values if TRUE.}

\item{show_cat_name}{show categories if TRUE.}

\item{show_serie_name}{show names of series if TRUE.}

\item{show_percent}{show percentages if TRUE.}

\item{separator}{separator for displayed labels.}
}
\description{
Data labels show details about data series. This function indicate that
data labels should be displayed. See \code{link{chart_labels_text}} for modifying
text settings associated with labels.
}
