% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data_color.R
\name{chart_data_smooth}
\alias{chart_data_smooth}
\title{Smooth series}
\usage{
chart_data_smooth(x, values)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{values}{`integer(num of series)`: a set of smooth values to map data values to.
It is a named vector, the values will be matched based on the names.
Possible values are 0 or 1
If it contains only one integer it will be associated to all existing series.}
}
\description{
Specify mappings from levels in the data to smooth or not lines.
}
\examples{
linec <- ms_linechart(data = iris, x = "Sepal.Length",
  y = "Sepal.Width", group = "Species")
linec <- chart_data_smooth(linec,
  values = c(virginica = 0, versicolor = 0, setosa = 0) )
}
\seealso{
\code{\link{chart_data_fill}}, \code{\link{chart_data_stroke}}, \code{\link{chart_data_size}}
}
