% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{chart_labels}
\alias{chart_labels}
\title{Modify axis and plot labels}
\usage{
chart_labels(x, title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{title, xlab, ylab}{Text to add}
}
\description{
Add labels to a chart, labels can be specified for
x axis, y axis and plot.
}
\examples{
mylc <- ms_linechart(data = browser_ts, x = "date", y = "freq",
  group = "browser")
mylc <- chart_labels(mylc, title = "my title", xlab = "my x label",
  ylab = "my y label")
}
