% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{ms_combochart}
\alias{ms_combochart}
\title{combochart object}
\usage{
ms_combochart(...)
}
\arguments{
\item{...}{mschart objects}
}
\description{
Creation of a combochart object that can be
inserted in a 'Microsoft' document.
}
\details{
ms_combochart only works with mschart objects created with
\code{asis = TRUE}.
}
\examples{
library(mschart)


# example chart 01 -------

dat <- longley
dat$Year <- as.Date(paste0(dat$Year, "-01-01"))
dat$`GNP Deflator` <- dat$GNP.deflator

empl <- dat |>
  ms_barchart(x = "Year", y = c("Employed", "Unemployed"), asis = TRUE) |>
  chart_labels(title = "Longley's Economic Regression Data",
               xlab = "1947 to 1962", ylab = "Employment") |>
  chart_ax_x(num_fmt = "yyyy") |>
  as_bar_stack(dir = "vertical")

gdp <- dat |>
  ms_linechart(x = "Year", y = "GNP Deflator", asis = TRUE) |>
  chart_labels(ylab = "GNP implicit price deflator (1954 = 100)") |>
  chart_ax_y(second_axis = TRUE)

cb <- ms_combochart(empl, gdp)
}
\seealso{
\code{\link[=chart_settings]{chart_settings()}}, \code{\link[=chart_ax_x]{chart_ax_x()}}, \code{\link[=chart_ax_y]{chart_ax_y()}},
\code{\link[=chart_data_labels]{chart_data_labels()}}, \code{\link[=chart_theme]{chart_theme()}}, \code{\link[=chart_labels]{chart_labels()}}

Other 'Office' chart objects: 
\code{\link{ms_areachart}()},
\code{\link{ms_barchart}()},
\code{\link{ms_linechart}()},
\code{\link{ms_piechart}()},
\code{\link{ms_scatterchart}()}
}
\concept{'Office' chart objects}
