% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{ms_combochart}
\alias{ms_combochart}
\title{combochart object}
\usage{
ms_combochart(...)
}
\arguments{
\item{...}{mschart objects}
}
\description{
Creation of a combochart object that can be
inserted in a 'Microsoft' document.
}
\details{
ms_combochart only works with mschart objects created with
\code{asis = TRUE}.
}
\seealso{
\code{\link[=chart_settings]{chart_settings()}}, \code{\link[=chart_ax_x]{chart_ax_x()}}, \code{\link[=chart_ax_y]{chart_ax_y()}},
\code{\link[=chart_data_labels]{chart_data_labels()}}, \code{\link[=chart_theme]{chart_theme()}}, \code{\link[=chart_labels]{chart_labels()}}

Other 'Office' chart objects: 
\code{\link{ms_areachart}()},
\code{\link{ms_barchart}()},
\code{\link{ms_linechart}()},
\code{\link{ms_piechart}()},
\code{\link{ms_scatterchart}()}
}
\concept{'Office' chart objects}
