% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{ms_piechart}
\alias{ms_piechart}
\title{piechart object}
\usage{
ms_piechart(data, x, y, group = NULL, labels = NULL, asis = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{x}{x colname}

\item{y}{y colname}

\item{group}{grouping colname used to split data into series. Optional.}

\item{labels}{colnames of columns to be used as labels into series. Optional.
If more than a name, only the first one will be used as label, but all
labels (transposed if a group is used) will be available in the Excel file
associated with the chart.}

\item{asis}{bool parameter defaulting to FALSE. If TRUE the data will not be
modified.}
}
\description{
Creation of a piechart object that can be
inserted in a 'Microsoft' document.
}
\seealso{
\code{\link[=chart_settings]{chart_settings()}}, \code{\link[=chart_ax_x]{chart_ax_x()}}, \code{\link[=chart_ax_y]{chart_ax_y()}},
\code{\link[=chart_data_labels]{chart_data_labels()}}, \code{\link[=chart_theme]{chart_theme()}}, \code{\link[=chart_labels]{chart_labels()}}

Other 'Office' chart objects: 
\code{\link{ms_areachart}()},
\code{\link{ms_barchart}()},
\code{\link{ms_linechart}()},
\code{\link{ms_scatterchart}()}
}
\concept{'Office' chart objects}
