% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mschart}
\alias{mschart}
\alias{mschart-package}
\title{mschart: Chart Generation for 'Microsoft Word' and 'Microsoft PowerPoint' Documents}
\description{
It lets R users to create Microsoft Office charts from data, and
then add title, legends, and annotations to the chart object.
}
\details{
The graph produced is a Microsoft graph, which means that it can be edited in
your Microsoft software and that the underlying data are available.

The package will not allow you to make the same charts as with ggplot2. It
allows only a subset of the charts possible with 'Office Chart'. The package is
often used to industrialize graphs that are then consumed and annotated by non-R
users.

The following charts are the only available from all possible MS charts:
\itemize{
\item barcharts: \code{\link[=ms_barchart]{ms_barchart()}}
\item line charts: \code{\link[=ms_linechart]{ms_linechart()}}
\item scatter plots: \code{\link[=ms_scatterchart]{ms_scatterchart()}}
\item area charts: \code{\link[=ms_areachart]{ms_areachart()}}
}
}
\seealso{
\url{https://ardata-fr.github.io/officeverse/}
}
\author{
\strong{Maintainer}: David Gohel \email{david.gohel@ardata.fr}

Other contributors:
\itemize{
  \item ArData [copyright holder]
  \item YouGov [funder]
  \item Rokas Klydzia (custom labels) [contributor]
  \item David Camposeco \email{david.camposeco.paulsen@gmail.com} (chart_data_smooth function) [contributor]
  \item Dan Joplin (fix scatter plot data structure) [contributor]
}

}
