% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R, R/mschart.R
\name{ms_linechart}
\alias{ms_linechart}
\alias{ms_barchart}
\alias{mschart}
\title{chart object}
\usage{
ms_linechart(data, x, y, group = NULL)

ms_barchart(data, x, y, group = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{x}{x colname}

\item{y}{y colname}

\item{group}{grouping colname used to split data into series. Optional.}
}
\description{
Creation of a chart object that can be
inserted in a 'Microsoft' document.
}
\section{Functions}{
\itemize{
\item \code{ms_linechart}: linechart function

\item \code{ms_barchart}: barchart function
}}

