% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_chart.R
\name{print.ms_chart}
\alias{print.ms_chart}
\title{Print method for ms_chart}
\usage{
\method{print}{ms_chart}(x, preview = FALSE, ...)
}
\arguments{
\item{x}{an \code{ms_chart} object.}

\item{preview}{preview the chart in a PowerPoint document}

\item{...}{unused}
}
\value{
No return value, called for side effects.
}
\description{
An \code{ms_chart} object cannot be rendered
in R. The default printing method will only display
simple information about the object.
If argument \code{preview} is set to TRUE, a \code{pptx} file
will be produced and opened with function \code{browseURL}.
}
