% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_indus_prod}
\alias{us_indus_prod}
\title{Index of US Industrial Production}
\format{
A data frame with 256 rows and 3 variables
}
\usage{
data(us_indus_prod)
}
\description{
Index of US industrial production (1985 = 100).
}
\details{
This is a transformation into simple data.frame
of data \code{USProdIndex} in package 'AER'.
}
\keyword{datasets}
